Executing a console command now requires 2 tmux commands.
This is necessary because tmux doesn't want to send the
word "end" to the bungee console. It may be because it's a
reserved keyword.

To fix this, I had no choice but to split the commands into
two: the first sends the command in a literal form. The
seccond command sends the Enter key.

I hope there is a better way to do this. For now, this will
suffice.